feat: upgrade to OTel SDK 2.0.0, extend support down to Node.js 16
Use OpenTelemetry JS SDK 2.x and matching related packages for Node.js
versions >= 18.19.0. Fall back to SDK 1.x and matching related packages
for Node.js versions >= 16.0.0 and < 18.19.0.

This release extends the range of supported versions down to Node.js
16.x (previously only Node.js >= 18.x had been supported).

BREAKING CHANGE: Due to the automatic fallback to SDK 1.x for older
Node.js versions, there is no actual breaking change in this release.
This comment just serves the purpose of forcing semantic-release to
publish this as version 2.0.0 instead of 1.2.0, which seems warranted
for the upgrade to OTel JS SDK 2.x. Strictly speaking, the fact that
Node.js 16.x processes will now be automatically instrumented as well
might be seen as a breaking change.